% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_standard_error}
\alias{calculate_standard_error}
\title{Calculate Standard Error}
\usage{
calculate_standard_error(ultimate_claims, cumulative_triangle)
}
\arguments{
\item{ultimate_claims}{A numeric vector of ultimate claims for each accident period.}

\item{cumulative_triangle}{A matrix or data frame representing the cumulative loss triangle.}
}
\value{
A numeric vector containing the standard error for each accident period.
}
\description{
Estimates the standard error for each accident period based on residuals.
}
\examples{
\dontrun{
data(RAA, package = "ChainLadder")
cumulative_triangle <- calculate_cumulative_triangle(RAA)
ultimate_claims <- rowSums(cumulative_triangle, na.rm = TRUE)
std_error <- calculate_standard_error(ultimate_claims, cumulative_triangle)
print(std_error)
}
}
