% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo_MackCP}
\alias{demo_MackCP}
\title{MackCP Demo Script}
\usage{
demo_MackCP()
}
\description{
This script demonstrates the key capabilities of the MackCP package using the RAA dataset.
It applies Mack's estimator motivated by large exposure asymptotics in a compound Poisson setting,
compares the results with the ChainLadder package's Mack estimator, and provides a detailed comparison.
}
\examples{
\dontrun{
demo_MackCP()
}
}
