% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mack_estimator.R
\name{print.mack_estimator}
\alias{print.mack_estimator}
\title{Print Method for Mack Estimator Results}
\usage{
\method{print}{mack_estimator}(x, ...)
}
\arguments{
\item{x}{An object of class 'mack_estimator'.}

\item{...}{Additional parameters.}
}
\description{
Print Method for Mack Estimator Results
}
