% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mack_estimator.R
\name{summary.mack_estimator}
\alias{summary.mack_estimator}
\title{Summary Method for Mack Estimator Results}
\usage{
\method{summary}{mack_estimator}(object, ...)
}
\arguments{
\item{object}{An object of class 'mack_estimator'.}

\item{...}{Additional parameters.}
}
\value{
A summary of the key results.
}
\description{
Summary Method for Mack Estimator Results
}
